+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.4.11
 - Removed debug log configuration option, as it's no longer needed (raoulvdberge)
 - Fixed bug where storages that are removed remain visible (raoulvdberge)
+- Added support for External Storage on Interfaces and other Refined Storage blocks, so you can keep items in stock easier (raoulvdberge)
 
 ### 1.4.10
 - Improved performance of network scanning (raoulvdberge)